feat: add loading cursor and disable buttons during async operations in DeletePopup, Elements, and CreateUpdateItem components

+++ web/app/(pages)/admin/components/DeletePopup.tsx
@@ -1,4 +1,5 @@
 "use client";
+import { useState } from "react";
 import config from "@/app/config-api.json";
 import styles from "./DeletePopup.module.css";
 
@@ -11,7 +12,12 @@ interface DeletePopupProps {
 };
 
 export default function DeletePopup({ actualSection, id, name, onClose, refetch }: DeletePopupProps) {
+  const [isDeleting, setIsDeleting] = useState(false);
+
   const handleDelete = async () => {
+    if (isDeleting) return null;
+    setIsDeleting(true);
+
     try {
       await fetch(`${config.api_url}/${actualSection}/variant/${id}`, {
         method: "DELETE"
@@ -22,10 +28,11 @@ export default function DeletePopup({ actualSection, id, name, onClose, refetch 
     }
     onClose();
     refetch();
+    setIsDeleting(false);
   }
 
   return (
-    <div className={styles.container}>
+    <div className={`${styles.container} ${isDeleting ? styles.cursorLoading : ""}`}>
 
       <p className={styles.text}>Tem certeza que gostaria de deletar: <br/> <strong className={styles.name}>{name}</strong>?</p>
 
@@ -35,8 +42,9 @@ export default function DeletePopup({ actualSection, id, name, onClose, refetch 
           onClick={onClose}
         ><p className={styles.textBtn}>Cancelar</p></button>
         <button
-          className={styles.delete}
+          className={isDeleting ? styles.disabled : styles.delete}
           onClick={handleDelete}
+          disabled={isDeleting}
         ><p className={styles.textBtn}>Deletar</p></button>
       </div>
     </div>
